requirements.txt
starlette
uvicorn
aiohttp
nest-asyncio~=1.6.0
python-dotenv~=1.1.0
prompts~=0.0.1
mcp~=1.9.2
